fic-2025-85-g13-sesion05/requirements.txt
Adafruit-Blinka==8.55.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-pn532==2.4.3
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.77.0
Adafruit-PN532==1.2.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
cffi==1.17.1
cryptography==44.0.2
lgpio==0.2.2.0
libusb1==3.3.0
mpu6050==0.0.3
ndeflib==0.3.3
nfcpy==1.0.4
pycparser==2.22
pyDes==2.0.1
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
quat==1.0.0
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
smbus==1.1.post2
spidev==3.6
sysv-ipc==1.1.0
typing_extensions==4.12.2
